Use slices for each part of the state to keep it more maintainable
